+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Do not edit the NEW_VERSION and VERSION_DATE fields, they will be updated automa
 
 ## [<NEW_VERSION>] - <VERSION_DATE>
 ### Added
+- Optional Google tag (gtag) output: user properties and custom event with ai_bot_status and ai_bot_is_bot, configurable measurement ID
 ### Changed
 ### Fixed